deprecate sunpy.physics.solar_rotate_coordinate and sunpy.physics.differential_rotate
