+++ sunpy/map/tests/reference/test_differential_rotation_maputil.txt
